requirements.txt
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
pytest==7.2.1

hydra-core==1.3.1